requirements.txt
alyn==0.1.1
autocorrect==0.3.0
backports.csv==1.0.6
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.3
chardet==3.0.4
cheroot==6.5.2
CherryPy==18.0.0
cloudpickle==0.5.5
contextlib2==0.5.5
cycler==0.10.0
dask==0.19.0
decorator==4.3.0
feedparser==5.2.1
future==0.16.0
idna==2.7
jaraco.functools==1.20
kiwisolver==1.0.1
lxml==4.2.4
matplotlib==2.2.3
more-itertools==4.3.0
networkx==2.1
nltk==3.3
numpy==1.15.1
opencv-python==3.4.2.17
Pattern==3.6
pdfminer.six==20170720
Pillow==5.2.0
pillowfight==0.3
portend==2.3
pycryptodome==3.6.6
pyparsing==2.2.0
pypillowfight==0.2.4
pytesseract==0.2.4
python-dateutil==2.7.3
python-docx==0.8.7
pytz==2018.5
PyWavelets
requests==2.19.1
scikit-image==0.14.0
scipy==1.1.0
six==1.11.0
tempora==1.13
tesserocr==2.3.1
toolz==0.9.0
urllib3==1.23
zc.lockfile==1.3.0
cachetools==2.1.0
certifi==2018.10.15
chardet==3.0.4
google-api-core==1.4.1
google-auth==1.5.1
google-cloud-speech==0.36.0
googleapis-common-protos==1.5.3
grpcio==1.15.0
idna==2.7
protobuf==3.6.1
pyasn1==0.4.4
pyasn1-modules==0.2.2
PyAudio==0.2.11
pydub==0.23.0
pytz==2018.5
requests==2.19.1
rsa==4.0
six==1.11.0
SpeechRecognition==3.8.1
urllib3==1.23
unidecode
